requirements.txt
asgiref==3.8.1
attrs==25.3.0
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
phonenumbers==9.0.7
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
